tests: fix a test
Commit 4b040c06c3e8aa3afc6c3051902fee399a258d05 inadvertently "fixed" an
issue (regression) that changed behavior from when this was a part of
Gentoo Build Publisher. Regarding inheritance, if a base class requires
(directly or indirectly) the same fixture as the child class, then the
fixture is only created once, from the parent's .setUp().  As a result,
the fixture is created using the parent's .options, not the child's.

The reason why the fixture is only created once is so, e.g., there won't
be differnt tests using different tmpdirs in the same (child) test. The
reason for using the parent's options is because any tests defined in
the parent class might be expecting those options and the child class
could override them.

This is perhaps not the obvious behavior and possibly doesn't always
"make sense". For now care should be taken when using inheritance where
base classes also use fixtures.