@requires: accept **kwargs
requires now accepts keyword arguments so that, e.g.:

@requires(this=that)
class MyTests(TestCase):
    ....

Whe MyTests.setUp is called. the fixture will be named "this" but the
fixture function is that.

This will allow you to use a fixture by a different name, or possibly
the same fixture function with different names.