State: remove fixtures attribute
This was needed when .setUp was overridden and set up the fixtures,
which is needed when the test method is called. However now that the
fixtures are set up in the test methods themselves this shared state is
no longer needed. However some functions had to be changed to accept and
return Fixtures as a result (no more side-effects).